+++ reference/rust/rrq-orchestrator/src/commands/mod.rs
@@ -1,0 +1,20 @@
+pub(crate) mod debug;
+pub(crate) mod dlq;
+pub(crate) mod executor;
+pub(crate) mod health;
+pub(crate) mod job;
+pub(crate) mod queue;
+pub(crate) mod shared;
+pub(crate) mod worker;
+
+pub(crate) use debug::{
+    debug_clear, debug_generate_jobs, debug_generate_workers, debug_stress_test, debug_submit,
+};
+pub(crate) use dlq::{
+    dlq_inspect, dlq_list, dlq_requeue, dlq_stats, DlqListOptions, DlqRequeueOptions,
+};
+pub(crate) use executor::executor_python;
+pub(crate) use health::check_workers;
+pub(crate) use job::{job_cancel, job_list, job_replay, job_show, job_trace};
+pub(crate) use queue::{queue_inspect, queue_list, queue_stats};
+pub(crate) use worker::{run_worker, run_worker_watch};
